.claude/skills/finance-process/pyproject.toml
[project]
name = "finance-process"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.7",
    "python-dateutil>=2.9.0.post0",
]
